+++ Stephen Wong CGP600 AE2/Game Files/Game/Lights/PointLight.h
@@ -1,7 +1,7 @@
 // *********************************************************
 //	Name:			Stephen Wong
 //	File:			PointLight.h
-//	Last Updated:	24/12/2017
+//	Last Updated:	03/01/2018
 //	Project:		CGP600 AE2
 // *********************************************************
 #pragma once
@@ -18,6 +18,6 @@ private:
 
 public:
 	PointLight(float red, float green, float blue,
-		float x, float y, float z, int range);
+		float x, float y, float z, float range);
 	XMVECTOR GetShineFromVector(XMMATRIX worldMatrix);
 };
